+++ NEWS.rst
@@ -11,9 +11,11 @@ AssignGenes:
 
 MakeDb:
 
-+ Added the ``-strict`` argument to multiple subcommands to provide more control
-  over the passing record validation criteria. ``--strict`` is mutually exclusive
-  with the ``--partial`` argument.
++ Modified the default validation criteria for passing records to exclude
+  validation that the junction region starts at position 310 in the
+  IMGT-numbered sequence. The new ``-strict`` argument is equivalent to the
+  prior behavior and is mutually exclusive with the ``--partial`` argument
+  (which is unchanged).
 + Added support for C region alignment calls provide by IgBLAST v1.18+.
 + Added the ``airr`` subcommand which will modify the ``sequence_alignment`` and
   ``germline_aligment`` fields in an AIRR Rearrangement file to include